casts for 64bit
git-svn-id: svn+ssh://svn.gna.org/svn/gnustep/libs/base/trunk@30738 72102866-910b-0410-8b05-ffd578937521

+++ Source/NSPropertyList.m
@@ -3498,11 +3498,11 @@ isEqualFunc(const void *item1, const void *item2,
 {
   int index;
 
-  index = (int)[objectList objectForKey: object];
+  index = (int)(intptr_t)[objectList objectForKey: object];
   if (index <= 0)
     {
       index = [objectList count];
-      [objectList setObject: (id)(++index) forKey: object];
+      [objectList setObject: (id)(intptr_t)(++index) forKey: object];
       [objectsToDoList addObject: object];
     }
